requirements.txt
# Core dependencies
openai>=1.0.0
networkx>=3.0
ipycytoscape>=1.3.1
ipywidgets>=8.0.0
pandas>=2.0.0

# Jupyter support
jupyter>=1.0.0
notebook>=7.0.0

# Configuration management
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Code quality
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Optional: For additional features
# neo4j>=5.0.0  # Graph database support
# spacy>=3.0.0  # Advanced NLP features